requirements.txt
flask
requests
beautifulsoup4
gunicorn
flask-cors
google-generativeai